refactoring: types placed in types folder and some clean up
