+++ src/components/tables/faculty/faculty-table.tsx
@@ -3,7 +3,7 @@ import { Button } from "~/components/ui/button";
 import { DataTable } from "~/components/ui/data-table";
 import { Heading } from "~/components/ui/heading";
 import { Separator } from "~/components/ui/separator";
-import { Faculty } from "~/constants/data";
+import { Faculty } from "~/types";
 import { Plus } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { columns } from "./columns";
